requirements.txt
click==8.1.8
docopt==0.6.2
Jinja2==3.1.5
jiwer==3.1.0
MarkupSafe==3.0.2
num2words==0.5.14
PyYAML==6.0.2
RapidFuzz==3.12.1
